requirements.txt
streamlit
google-genai
python-dotenv
opencv-python-headless
pillow
pydantic
datasets
evaluate
kaggle
google-generativeai
